requirements.txt
Django>=2.1.4
whitenoise>=4.1.2
gunicorn>=19.9.0
Pillow>=5.3.0
dj-database-url>=0.5.0
mysqlclient>=1.3.12
factory-boy>=2.11.1